# backend/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tingyue Backend API",
    "version": "1.0.0",
    "description": "Authoritative API contract api-contract-v1.0.0. JSON fields use camelCase and timestamps use UTC RFC 3339. Clients branch on stable error.code values, never human messages."
  },
  "servers": [
    { "url": "http://127.0.0.1:3100", "description": "Local development" },
    { "url": "https://api.test.example.invalid", "description": "Test placeholder; replace after domain approval" }
  ],
...
  "security": [{ "bearerAuth": [] }],
  "paths": {
...
    "/api/v1/me/quiz-attempts": {
      "post": {
        "operationId": "submitQuizAttempt",
        "tags": ["Quiz"],
        "description": "Idempotency identity is userId plus attemptId. Client-provided score, mastery and ranking are forbidden by the request schema.",
        "parameters": [{ "$ref": "#/components/parameters/IdempotencyKey" }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "./schemas/api-v1.schema.json#/$defs/QuizAttemptRequest" } } }
        },
        "responses": {
          "200": {
            "description": "Server-verified score or stable rejected result. Exact duplicate attempts return the original result.",
            "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
            "content": { "application/json": { "schema": { "$ref": "./schemas/api-v1.schema.json#/$defs/QuizAttemptResponse" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "401": { "$ref": "#/components/responses/Unauthorized" },
          "409": { "$ref": "#/components/responses/Conflict" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Short-lived access token. Tokens must not be logged."
      }
    },
    "headers": {
      "RequestId": {
        "description": "Opaque request correlation identifier.",
        "required": true,
        "schema": { "type": "string", "minLength": 8, "maxLength": 128 }
      }
    },
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": true,
        "description": "Opaque key unique to one logical mutation. Reusing it with a different request body returns IDEMPOTENCY_KEY_REUSED.",
        "schema": { "type": "string", "minLength": 8, "maxLength": 128 }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Malformed request, invalid ranking period or invalid filter.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "./schemas/api-v1.schema.json#/$defs/ErrorResponse" } } }
      },
      "Unauthorized": {
        "description": "Authentication failed, expired or was revoked.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "./schemas/api-v1.schema.json#/$defs/ErrorResponse" } } }
      },
...
      "Conflict": {
        "description": "State, version or idempotency conflict.",
        "headers": { "X-Request-Id": { "$ref": "#/components/headers/RequestId" } },
        "content": { "application/json": { "schema": { "$ref": "./schemas/api-v1.schema.json#/$defs/ErrorResponse" } } }
      },
...
    }
  }
}